Refactor introspection into modules
