+++ dbus.cabal
@@ -110,7 +110,9 @@ library
     DBus.Internal.Message
     DBus.Internal.Types
     DBus.Internal.Wire
-    DBus.Introspection
+    DBus.Introspection.Parse
+    DBus.Introspection.Render
+    DBus.Introspection.Types
     DBus.Socket
     DBus.TH
     DBus.Transport
