+++ examples/introspect.hs
@@ -24,7 +24,8 @@ import           System.IO (hPutStrLn, stderr)
 
 import           DBus
 import           DBus.Client
-import qualified DBus.Introspection as I
+import qualified DBus.Introspection.Parse as I
+import qualified DBus.Introspection.Types as I
 
 main :: IO ()
 main = do
